+++ src/test/com/extrahardmode/service/BlockTypeTest.java
@@ -31,13 +31,13 @@ public class BlockTypeTest
     @Test
     public void matches_advanced()
     {
-        BlockType block = new BlockType(3, (byte) 1, (byte) 3, (byte) 4);
+        BlockType block = new BlockType(3, (short) 1, (short) 3, (short) 4);
 
-        assertTrue(block.matches(3, (byte) 1));
-        assertFalse(block.matches(3, (byte) 2));
-        assertTrue(block.matches(3, (byte) 4));
+        assertTrue(block.matches(3, (short) 1));
+        assertFalse(block.matches(3, (short) 2));
+        assertTrue(block.matches(3, (short) 4));
 
-        assertFalse(block.matches(1, (byte) 1));
+        assertFalse(block.matches(1, (short) 1));
     }
 
 
@@ -63,7 +63,7 @@ public class BlockTypeTest
     public void loadFromConfig_meta_simple()
     {
         String input = "Stone,2";
-        BlockType expected = new BlockType(Material.STONE, (byte) 2);
+        BlockType expected = new BlockType(Material.STONE, (short) 2);
         assertEquals(expected, BlockType.loadFromConfig(input));
     }
 
@@ -72,7 +72,7 @@ public class BlockTypeTest
     public void loadFromConfig_meta_advanced()
     {
         String input = "Stone,2,3,4";
-        BlockType expected = new BlockType(Material.STONE, (byte) 2, (byte) 3, (byte) 4);
+        BlockType expected = new BlockType(Material.STONE, (short) 2, (short) 3, (short) 4);
         assertEquals(expected, BlockType.loadFromConfig(input));
     }
 
@@ -81,7 +81,7 @@ public class BlockTypeTest
     public void saveToConfig_simple()
     {
         String expected = "STONE@1,2,3";
-        BlockType block = new BlockType(Material.STONE, (byte) 1, (byte) 2, (byte) 3);
+        BlockType block = new BlockType(Material.STONE, (short) 1, (short) 2, (short) 3);
         assertEquals(expected, block.saveToString());
     }
 }
